refactor(loss_hub): rename logits.py -> logit_processors.py
The module contains `get_responses`, `get_log_probs_and_entropy`, and
`get_values` — all functions that *process* model logits to extract
training-time signals. `logit_processors` better names the role; the
old `logits` collided in meaning with the tensor argument passed to
those functions.
